app/docs/public-api-examples/python/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "si-public-api-examples-python"
version = "1.0.0"
description = "Test coverage for System Initiative Python SDK"
authors = [
    {name = "System Initiative", email = "support@systeminit.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
dependencies = [
    "system-initiative-api-client>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.setuptools.packages.find]
exclude = ["tests*"]
